This patch replaces most instances of PRIVMSG with NOTICE.
According to RFC 2812, Section 3.3.2, PRIVMSG is for messages that other bots should reply to. NOTICE is the same, but other bots should not reply. If oiler is in a channel with another bot, the !quote function can be abused to issue bot command to other bots that respond to PRIVMSG. Vulnerabilities resulting from this include double voting and infinite loops with other quoting bots.

<http://tools.ietf.org/html/rfc2812#section-3.3.2>